requirements.txt
Django>1.10
djangorestframework>3.3
six
pytz
mkdocs-material

